+++ polaris-react/scripts/utilities/index.js
@@ -1,3 +1,0 @@
-const {semverRegExp} = require('./semver-reg-exp');
-
-module.exports.semverRegExp = semverRegExp;