+++ polaris.shopify.com/pages/examples/index-table-default.tsx
@@ -4,7 +4,6 @@ import {
   useIndexResourceState,
   Text,
   Badge,
-  useBreakpoints,
 } from '@shopify/polaris';
 import React from 'react';
 import {withPolarisExample} from '../../src/components/PolarisExampleWrapper';
@@ -79,7 +78,6 @@ function SimpleIndexTableExample() {
   return (
     <LegacyCard>
       <IndexTable
-        condensed={useBreakpoints().smDown}
         resourceName={resourceName}
         itemCount={orders.length}
         selectedItemsCount={